Move code to subfolder for samlproviderconfig
