+++ common/djangoapps/third_party_auth/samlproviderconfig/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from .models import SAMLProviderConfig
+from third_party_auth.models import SAMLProviderConfig
 
 class SAMLProviderConfigSerializer(serializers.ModelSerializer):
     class Meta:
